successfully introspecting and parsing input type
