+++ todo/module/.gitignore
@@ -105,4 +105,4 @@ dist
 
 # react-relay-app 
 # /react-relay-app/vanilla-gql-example.js
-../react-relay-app/__generated__
+../../react-relay-app/__generated__
